Code cleanup. Added custom helper functions to uiview to make programmatic constraints easier to manage.
