+++ Drip/Util/Resources/Custom Views/DrinksLauncher.swift
@@ -1,6 +1,6 @@
 import UIKit
 
-class DrinksLauncher: NSObject, UICollectionViewDataSource, UICollectionViewDelegate, UICollectionViewDelegateFlowLayout {
+class DrinksLauncher: NSObject {
 
     let blackView = UIView()
 
@@ -47,55 +47,9 @@ class DrinksLauncher: NSObject, UICollectionViewDataSource, UICollectionViewDele
 
         if let window = UIApplication.shared.windows.filter({$0.isKeyWindow}).first {
 
-            // black background view
-            blackView.backgroundColor = UIColor(white: 0, alpha: 0.5)
-            blackView.addGestureRecognizer(UITapGestureRecognizer(target: self, action: #selector(handleDismiss)))
-
-            containerview.backgroundColor = .purple
-            containerview.layer.cornerRadius = 20
-            containerview.layer.maskedCorners = [.layerMinXMinYCorner, .layerMaxXMinYCorner]
-
             window.addSubview(blackView)
             window.addSubview(containerview)
             print(window.subviews.count)
-
-            containerview.addSubview(titleLabel)
-            containerview.addSubview(buttonsContainer)
-            containerview.addSubview(collectionView)
-            containerview.addSubview(pageControl)
-
-            titleLabel.centerHorizontallyInSuperview()
-            titleLabel.anchor(top: containerview.topAnchor,
-                              leading: nil, bottom: nil, trailing: nil,
-                              padding: .init(top: 10, left: 0, bottom: 0, right: 0))
-
-            buttonsContainer.backgroundColor = .gray
-            buttonsContainer.layer.cornerRadius = 5
-            buttonsContainer.anchor(top: titleLabel.bottomAnchor,
-                                    leading: containerview.leadingAnchor,
-                                    bottom: nil,
-                                    trailing: containerview.trailingAnchor,
-                                    padding: .init(top: 10, left: 10, bottom: 0, right: 10),
-                                    size: .init(width: 0, height: 100))
-
-            buttonsContainer.addSubview(horizontalSV)
-            horizontalSV.anchor(top: nil,
-                                leading: buttonsContainer.leadingAnchor,
-                                bottom: nil,
-                                trailing: buttonsContainer.trailingAnchor,
-                                padding: .zero, size: CGSize(width: buttonsContainer.frame.width,
-                                                             height: 80))
-            horizontalSV.centerVerticallyInSuperview()
-
-            collectionView.anchor(top: buttonsContainer.bottomAnchor,
-                                  leading: containerview.leadingAnchor,
-                                  bottom: nil,
-                                  trailing: containerview.trailingAnchor,
-                                  padding: .init(top: 10, left: 10, bottom: 0, right: 10),
-                                  size: .init(width: 0, height: 260))
-            pageControl.anchor(top: collectionView.bottomAnchor,
-                               leading: nil, bottom: nil, trailing: nil, padding: .zero)
-            pageControl.centerHorizontallyInSuperview()
 
             blackView.frame = window.frame
             blackView.alpha = 0
@@ -115,6 +69,55 @@ class DrinksLauncher: NSObject, UICollectionViewDataSource, UICollectionViewDele
             }, completion: nil)
 
         }
+
+    }
+
+    func setupViews() {
+        // black background view
+        blackView.backgroundColor = UIColor(white: 0, alpha: 0.5)
+        blackView.addGestureRecognizer(UITapGestureRecognizer(target: self, action: #selector(handleDismiss)))
+
+        containerview.backgroundColor = .purple
+        containerview.layer.cornerRadius = 20
+        containerview.layer.maskedCorners = [.layerMinXMinYCorner, .layerMaxXMinYCorner]
+
+        containerview.addSubview(titleLabel)
+        containerview.addSubview(buttonsContainer)
+        containerview.addSubview(collectionView)
+        containerview.addSubview(pageControl)
+
+        titleLabel.centerHorizontallyInSuperview()
+        titleLabel.anchor(top: containerview.topAnchor,
+                          leading: nil, bottom: nil, trailing: nil,
+                          padding: .init(top: 10, left: 0, bottom: 0, right: 0))
+
+        buttonsContainer.backgroundColor = .gray
+        buttonsContainer.layer.cornerRadius = 5
+        buttonsContainer.anchor(top: titleLabel.bottomAnchor,
+                                leading: containerview.leadingAnchor,
+                                bottom: nil,
+                                trailing: containerview.trailingAnchor,
+                                padding: .init(top: 10, left: 10, bottom: 0, right: 10),
+                                size: .init(width: 0, height: 100))
+
+        buttonsContainer.addSubview(horizontalSV)
+        horizontalSV.anchor(top: nil,
+                            leading: buttonsContainer.leadingAnchor,
+                            bottom: nil,
+                            trailing: buttonsContainer.trailingAnchor,
+                            padding: .zero, size: CGSize(width: buttonsContainer.frame.width,
+                                                         height: 80))
+        horizontalSV.centerVerticallyInSuperview()
+
+        collectionView.anchor(top: buttonsContainer.bottomAnchor,
+                              leading: containerview.leadingAnchor,
+                              bottom: nil,
+                              trailing: containerview.trailingAnchor,
+                              padding: .init(top: 10, left: 10, bottom: 0, right: 10),
+                              size: .init(width: 0, height: 260))
+        pageControl.anchor(top: collectionView.bottomAnchor,
+                           leading: nil, bottom: nil, trailing: nil, padding: .zero)
+        pageControl.centerHorizontallyInSuperview()
 
     }
 
@@ -164,6 +167,25 @@ class DrinksLauncher: NSObject, UICollectionViewDataSource, UICollectionViewDele
 
     }
 
+    override init() {
+        super.init()
+        // do stuff
+        collectionView.dataSource = self
+        collectionView.delegate = self
+        collectionView.isPagingEnabled = true
+        collectionView.showsHorizontalScrollIndicator = false
+        pageControl.isUserInteractionEnabled = false
+
+        collectionView.register(DrinksCell.self, forCellWithReuseIdentifier: cellId)
+        setupViews()
+        setupStackView()
+
+    }
+
+}
+
+extension DrinksLauncher: UICollectionViewDataSource, UICollectionViewDelegate, UICollectionViewDelegateFlowLayout {
+
     func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
         let count = 18
         pageControl.numberOfPages = Int(ceil(Double(count)/6))
@@ -180,15 +202,27 @@ class DrinksLauncher: NSObject, UICollectionViewDataSource, UICollectionViewDele
         pageControl.currentPage = Int(scrollView.contentOffset.x) / Int(scrollView.frame.width)
     }
 
+    func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
+        print("item \(indexPath.item)")
+    }
+
     func collectionView(_ collectionView: UICollectionView,
                         cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        let cell = collectionView.dequeueReusableCell(withReuseIdentifier: cellId, for: indexPath) as! DrinksCell
-        cell.nameLabel.text = "test\(indexPath.item)"
 
-        let colours: [UIColor] = [.blue, .red, .red, .blue, .blue, .red, .red, .blue, .blue,
-                                  .red, .red, .blue, .blue, .red, .red,
-                                  .blue, .blue, .red, .red, .blue, .blue]
-        cell.backgroundColor = colours[indexPath.item]
+        var cell = UICollectionViewCell()
+
+        if let drinkCell = collectionView.dequeueReusableCell(withReuseIdentifier: cellId,
+                                                              for: indexPath) as? DrinksCell {
+            drinkCell.nameLabel.text = "test\(indexPath.item)"
+
+            let colours: [UIColor] = [.blue, .red, .red, .blue, .blue, .red, .red, .blue, .blue,
+                                      .red, .red, .blue, .blue, .red, .red,
+                                      .blue, .blue, .red, .red, .blue, .blue]
+            drinkCell.backgroundColor = colours[indexPath.item]
+
+            cell = drinkCell
+        }
+
         return cell
     }
 
@@ -205,85 +239,5 @@ class DrinksLauncher: NSObject, UICollectionViewDataSource, UICollectionViewDele
     func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout,
                         sizeForItemAt indexPath: IndexPath) -> CGSize {
         return CGSize(width: collectionView.bounds.width/3, height: collectionView.bounds.height/2)
-    }
-
-    override init() {
-        super.init()
-        // do stuff
-        collectionView.dataSource = self
-        collectionView.delegate = self
-        collectionView.isPagingEnabled = true
-        collectionView.showsHorizontalScrollIndicator = false
-        pageControl.isUserInteractionEnabled = false
-
-        collectionView.register(DrinksCell.self, forCellWithReuseIdentifier: cellId)
-        setupStackView()
-
-    }
-
-}
-
-extension UIView {
-
-    func anchor(top: NSLayoutYAxisAnchor?,
-                leading: NSLayoutXAxisAnchor?,
-                bottom: NSLayoutYAxisAnchor?,
-                trailing: NSLayoutXAxisAnchor?,
-                padding: UIEdgeInsets = .zero,
-                size: CGSize = .zero) {
-
-        translatesAutoresizingMaskIntoConstraints = false
-
-        if let top = top {
-            topAnchor.constraint(equalTo: top, constant: padding.top).isActive = true
-        }
-
-        if let leading = leading {
-            leadingAnchor.constraint(equalTo: leading, constant: padding.left).isActive = true
-        }
-
-        if let bottom = bottom {
-            bottomAnchor.constraint(equalTo: bottom, constant: -padding.bottom).isActive = true
-        }
-
-        if let trailing = trailing {
-            trailingAnchor.constraint(equalTo: trailing, constant: -padding.right).isActive = true
-        }
-
-        if size.width != 0 {
-            widthAnchor.constraint(equalToConstant: size.width).isActive = true
-        }
-
-        if size.height != 0 {
-            heightAnchor.constraint(equalToConstant: size.height).isActive = true
-        }
-    }
-
-    func centerInSuperview() {
-        if let superview = superview {
-            centerXAnchor.constraint(equalTo: superview.centerXAnchor).isActive = true
-            centerYAnchor.constraint(equalTo: superview.centerYAnchor).isActive = true
-        }
-    }
-
-    func centerHorizontallyInSuperview() {
-        if let superview = superview {
-            centerXAnchor.constraint(equalTo: superview.centerXAnchor).isActive = true
-        }
-    }
-
-    func centerVerticallyInSuperview() {
-        if let superview = superview {
-            centerYAnchor.constraint(equalTo: superview.centerYAnchor).isActive = true
-        }
-    }
-
-    func fillSuperView() {
-        if let superview = superview {
-            trailingAnchor.constraint(equalTo: superview.trailingAnchor).isActive = true
-            topAnchor.constraint(equalTo: superview.topAnchor).isActive = true
-            leadingAnchor.constraint(equalTo: superview.leadingAnchor).isActive = true
-            bottomAnchor.constraint(equalTo: superview.bottomAnchor).isActive = true
-        }
     }
 }
