+++ Drip/Util/Resources/Extensions/UIView+Extensions.swift
@@ -1,0 +1,75 @@
+//
+//  UIView+Extensions.swift
+//  Drip
+//
+//  Created by Kaiman Mehmet on 20/03/2021.
+//  Copyright © 2021 Kaiman Mehmet. All rights reserved.
+//
+
+import UIKit
+
+extension UIView {
+
+    func anchor(top: NSLayoutYAxisAnchor?,
+                leading: NSLayoutXAxisAnchor?,
+                bottom: NSLayoutYAxisAnchor?,
+                trailing: NSLayoutXAxisAnchor?,
+                padding: UIEdgeInsets = .zero,
+                size: CGSize = .zero) {
+
+        translatesAutoresizingMaskIntoConstraints = false
+
+        if let top = top {
+            topAnchor.constraint(equalTo: top, constant: padding.top).isActive = true
+        }
+
+        if let leading = leading {
+            leadingAnchor.constraint(equalTo: leading, constant: padding.left).isActive = true
+        }
+
+        if let bottom = bottom {
+            bottomAnchor.constraint(equalTo: bottom, constant: -padding.bottom).isActive = true
+        }
+
+        if let trailing = trailing {
+            trailingAnchor.constraint(equalTo: trailing, constant: -padding.right).isActive = true
+        }
+
+        if size.width != 0 {
+            widthAnchor.constraint(equalToConstant: size.width).isActive = true
+        }
+
+        if size.height != 0 {
+            heightAnchor.constraint(equalToConstant: size.height).isActive = true
+        }
+    }
+
+    func centerInSuperview() {
+        if let superview = superview {
+            centerXAnchor.constraint(equalTo: superview.centerXAnchor).isActive = true
+            centerYAnchor.constraint(equalTo: superview.centerYAnchor).isActive = true
+        }
+    }
+
+    func centerHorizontallyInSuperview() {
+        if let superview = superview {
+            centerXAnchor.constraint(equalTo: superview.centerXAnchor).isActive = true
+        }
+    }
+
+    func centerVerticallyInSuperview() {
+        if let superview = superview {
+            centerYAnchor.constraint(equalTo: superview.centerYAnchor).isActive = true
+        }
+    }
+
+    func fillSuperView() {
+        if let superview = superview {
+            trailingAnchor.constraint(equalTo: superview.trailingAnchor).isActive = true
+            topAnchor.constraint(equalTo: superview.topAnchor).isActive = true
+            leadingAnchor.constraint(equalTo: superview.leadingAnchor).isActive = true
+            bottomAnchor.constraint(equalTo: superview.bottomAnchor).isActive = true
+        }
+    }
+}
+
